ora_exception_lib/pyproject.toml
[project]
name = "ora-exception-lib"
version = "0.1.0"
description = "Custom exception handling library for FastAPI (Java-to-Python equivalent)"
readme = "README.md"
authors = [
  { name = "Siddharth", email = "siddharths.chaudhari@fosteringlinux.com" }
]

requires-python = ">=3.9"

dependencies = [
    "fastapi",
    "pydantic",
    "PyYAML"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
